+++ CHANGES.md
@@ -10,7 +10,42 @@
     - バグ修正
 
 
-## develop
+## 1.10.0
+
+### UPDATE
+
+- `minSdkVersion` を 21 に上げた
+  - `com.squareup.okhttp3:okhttp` 4.2.2 が `minSdkVersion` 21 以上にのみ対応するため
+- libwebrtc を 78.8.0 に上げた
+- Android Studio 3.5.1 に対応した
+- Kotlin を 1.3.50 に上げた
+- Dokka を 0.10.0 に上げた
+- `com.android.tools.build:gradle` を 3.5.2 に上げた
+- `com.squareup.okhttp3:okhttp` を 4.2.2 に上げた
+- `com.google.code.gson:gson` を 2.8.6 に上げた
+- `org.robolectric:robolectric` を 4.3.1 に上げた
+- AudioDeviceManager 生成時のパラメータをオプション `SoraAudioOption` に追加した
+  - `audioSource`: `android.media.MediaRecorder.AudioSource` のいずれか
+  - `useStereoInput`: boolean
+  - `useStereoOutput`: boolean
+
+### ADD
+
+- シグナリング connect メッセージに `sdk_type`, `sdk_version` と `user_agent` を追加した
+- シグナリング connect メッセージに `audio.opus_params` を追加した
+- 1:N サイマルキャストの視聴に対応した
+
+### CHANGE
+
+- 時雨堂ビルドの libwebrtc ライブラリ名称を変更した
+  - 旧: `sora-webrtc-android` 、 新: `shiguredo-webrtc-android`
+  - `transitive = true` で `sora-android-sdk` に依存している場合はアプリ側の変更は不要
+- シグナリング connect メッセージから `simulcast_rid` を削除した
+
+### FIX
+
+- 視聴のみかつ H.264 を指定した場合に接続できない現象を修正した
+
 
 ## 1.9.0
 
@@ -51,7 +86,7 @@
   - このバージョンでは `JavaAudioDeviceModule` の audio record, audio track 関連のエラーが
     このコールバックを通して通知される
 - rid-based simulcast に部分的に対応した
-  - 現状では、ソフトウェアエンコーダのみで動作する
+  - 現状では、ソフトウェアエンコーダの配信のみで動作する
   - 映像コーデックは VP8 のみの対応する
   - fixed resolution と一緒に使うとクラッシュ(SEGV)することが分かっている
     - 関連してそうな issue: 10713 - Transceiver/encodings based simulcast does not work in desktop sharing